+++ sdk/cosmosdb/cosmos/test/public/functional/container.spec.ts
@@ -487,8 +487,23 @@ describe("container.create", function () {
 });
 
 describe("container.deleteAllItemsForPartitionKey", function () {
-  it("should delete all items for partition key", async function () {
+  it("should delete all items for partition key value", async function () {
     const container = await getTestContainer("container", undefined, { partitionKey: "/pk" });
+    await testDeleteAllItemsForPartitionKey(container);
+  });
+
+  it("should delete all items for parition key value in multi partition container", async function () {
+    //multi partition container
+    const container = await getTestContainer("container", undefined, {
+      partitionKey: {
+        paths: ["/pk"],
+        version: 2,
+      },
+      throughput: 10500,
+    });
+    await testDeleteAllItemsForPartitionKey(container);
+  });
+  async function testDeleteAllItemsForPartitionKey(container: Container) {
     const { resource: create1 } = await container.items.create({
       id: "1",
       key: "value",
@@ -508,5 +523,5 @@ describe("container.deleteAllItemsForPartitionKey", function () {
     assert((await container.item(create1.id).read()).statusCode === StatusCodes.NotFound);
     assert((await container.item(create2.id).read()).statusCode === StatusCodes.NotFound);
     assert((await (await container.item(create3.id).read()).item.id) === create3.id);
-  });
+  }
 });